third_party/WebKit/LayoutTests/fast/dom/node-iterator-reference-node-removed-expected.txt

This test removes the NodeItertor's current reference node, then continues to move through the document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS iter.nextNode() is testDiv
PASS iter.nextNode() is div1
PASS iter.nextNode() is div2
PASS iter.nextNode() is null
PASS iter.previousNode() is div2
PASS iter.previousNode() is div1
PASS iter.nextNode() is div2
PASS successfullyParsed is true

TEST COMPLETE
